+++ pkg/resource/logging-agents-toggle/observability_bundle_configmap.go
@@ -1,4 +1,4 @@
-package appstoggle
+package loggingagentstoggle
 
 import (
 	"github.com/blang/semver"