+++ pkg/resource/logging-agents-toggle/reconciler.go
@@ -1,4 +1,4 @@
-package appstoggle
+package loggingagentstoggle
 
 import (
 	"context"