+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Configure grafana-agent config.
+- Configure grafana-agent config to grab Kubernetes Events and send them to Loki.
 
 ## [0.1.3] - 2023-10-18